Smaller types to handle overflow

+++ src/png.rs
@@ -243,10 +243,10 @@ impl PngFile {
             let mut i = 0;
             while i < self.pitch {
                 let x = pixel_start + i;
-                let p = pixels[x];
-                let a = if x - pixel_start > self.bytes_per_pixel - 1 { pixels[x - self.bytes_per_pixel] } else { 0 };
-                let b = if y > 0 { pixels[x - row_size] } else { 0 };
-                let c = if x - pixel_start > self.bytes_per_pixel - 1 && y > 0 { pixels[x - row_size - self.bytes_per_pixel] } else { 0 };
+                let p = pixels[x] as u16;
+                let a = if x - pixel_start > self.bytes_per_pixel - 1 { pixels[x - self.bytes_per_pixel] as u16 } else { 0 };
+                let b = if y > 0 { pixels[x - row_size] as u16 } else { 0 };
+                let c = if x - pixel_start > self.bytes_per_pixel - 1 && y > 0 { pixels[x - row_size - self.bytes_per_pixel] as u16 } else { 0 };
 
                 pixels[x] = filter.apply(p, a, b, c);
